add tests and fix query_filter
query_filter.query_with_filters() was failing to assign the return value of
f.apply() back to query, so the filters were not being applied in the loop,
resulting the results from the unfiltered query being returned no-matter what
queries are applied. Fixed it by assigning output back to query.