add InFilter validation that value is an iterable
the InFilter class validates that its value is an iterable, because it doesn't make sense to have an
in-clause on a scalar. Also made a quick optimization in the _get_filter_class and validate_operator
methods of the filter schema module to move their operator->filter map and valid operator set into
module level variables so they aren't executed every time those functions are called.